tests/wpt/tests/html/semantics/grouping-content/the-div-element/grouping-div.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>The div element</title>
    <link rel="author" title="dzenana" href="mailto:dzenana.trenutak@gmail.com">
    <link rel="help" href="https://html.spec.whatwg.org/multipage/#the-div-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
        "use strict";

        // check that prototype matches spec's DOM interface
        test(function () {
            var testElement = document.createElement("div");
            assert_equals(Object.getPrototypeOf(testElement), HTMLDivElement.prototype, "HTMLDivElement.prototype should be used for div");
        }, "The prototype for div is HTMLDivElement.prototype");

    </script>
</head>
<body>
    <h1>Description</h1>
    <p>This test validates the div element.</p>

    <div id="log"></div>

</body>
</html>
